pyproject.toml
[tool.poetry]
name = "douili"
version = "0.1.0"
description = ""
authors = ["diodonfrost <diodon.frost@diodonfrost.me>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.3.3"
python-dotenv = "^0.12.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
